// src/app/models/movie.dto.ts
export class Movie {
	adult: boolean;
	backdrop_path: string;
	genre_ids: number[];
	id: number;
	media_type: string;
	original_language: string;
	original_title: string;
	overview: string;
	popularity: number;
	poster_path: string;
	release_date: string;
	title: string;
	video: boolean;
	vote_average: number;
	vote_count: number;
	budget: number;
	homepage: string;
	imdb_id: number;
	runtime: number;
	genres: any[];
	spoken_languages: any[];
}